+++ backend/smody/src/main/java/com/woowacourse/smody/exception/dto/IssueCreateRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.smody.exception;
+package com.woowacourse.smody.exception.dto;
 
 import java.util.List;
